Re-organize cpp files for assembly methods.
